+++ simple_network_sim/common.py
@@ -1,0 +1,10 @@
+# CurrentlyInUse
+def generateMeanPlot(listOfPlots):
+    meanForPlot = []
+    # print(listOfPlots)
+    for i in range(len(listOfPlots[0])):
+        sumTot = 0
+        for j in range(len(listOfPlots)):
+            sumTot = sumTot + listOfPlots[j][i]
+        meanForPlot.append(float(sumTot)/len(listOfPlots))
+    return meanForPlot
